+++ app/src/main/java/com/theveloper/pixelplay/presentation/navigation/AppNavigation.kt
@@ -141,11 +141,6 @@ fun AppNavigation(
                     )
                 }
             }
-            composable(Screen.SongRecognition.route) {
-                ScreenWrapper(navController = navController, playerViewModel = playerViewModel, animatedVisibilityScope = this) {
-                    com.theveloper.pixelplay.presentation.screens.SongRecognitionScreen(paddingValues, playerViewModel)
-                }
-            }
             composable(
                 Screen.Search.route,
                 enterTransition = {